🐛 fix: use a double scroll offset in footer test refs #167

+++ test/shared/presentation/app_footer_host_test.dart
@@ -47,8 +47,9 @@ void main() {
 
     expect(find.byType(AppFooter), findsOneWidget);
 
-    final awayFromEnd =
-        (scrollController.position.maxScrollExtent - 100).clamp(0.0, double.infinity);
+    final awayFromEnd = (scrollController.position.maxScrollExtent - 100)
+        .clamp(0.0, double.infinity)
+        .toDouble();
     scrollController.jumpTo(awayFromEnd);
     await tester.pump();
     await tester.pump();
